+++ deployments/puppet/lib/facter/win_programdata.rb
@@ -1,7 +1,7 @@
 # Returns the PROGRAMDATA env var on windows
 
 Facter.add(:win_programdata) do
-  confine :osfamily => :windows
+  confine :kernel => 'windows'
   setcode do
     ENV['PROGRAMDATA']
   end