+++ deployments/puppet/lib/facter/win_systemdrive.rb
@@ -1,7 +1,7 @@
 # Returns the SystemDrive env var on windows
 
 Facter.add(:win_systemdrive) do
-  confine :osfamily => :windows
+  confine :kernel => 'windows'
   setcode do
     ENV['SystemDrive']
   end