+++ deployments/puppet/lib/facter/win_temp.rb
@@ -1,7 +1,7 @@
 # Returns the TEMP env var on windows
 
 Facter.add(:win_temp) do
-  confine :osfamily => :windows
+  confine :kernel => 'windows'
   setcode do
     ENV['TEMP']
   end